split into bin and lib
