+++ Cargo.toml
@@ -3,6 +3,10 @@ name = "mdproof"
 version = "0.1.0"
 authors = ["LeRoyce Pearson <leroycepearson@geemili.xyz>"]
 
+[[bin]]
+name = "mdproof"
+doc = false
+
 [dependencies]
 quicli = "0.3"
 pulldown-cmark = "0"
