+++ src/lib.rs
@@ -1,12 +1,21 @@
-#[macro_use] extern crate quicli;
 extern crate pulldown_cmark as cmark;
 extern crate pdf_canvas;
+
+mod span;
+mod section;
+mod sectioner;
+mod page;
+mod pages;
 
 use pdf_canvas::{Pdf, BuiltinFont};
 use cmark::*;
 use std::fs::File;
 use std::io::Read;
-use quicli::prelude::*;
+use std::io::Result;
+
+use span::Span;
+use sectioner::Sectioner;
+use pages::Pages;
 
 /// PAGE_SIZE is the size of a sheet of A4 paper in pt
 const PAGE_SIZE: (f32, f32) = (595.0, 842.0);
@@ -25,32 +34,10 @@ const LINE_SPACING: f32 = 1.75; // Text height * LINE_SPACING
 const LIST_INDENTATION: f32 = 20.0;
 const QUOTE_INDENTATION: f32 = 20.0;
 
-mod span;
-mod section;
-mod sectioner;
-mod page;
-mod pages;
+pub fn run(output_file: &str, markdown_file: &str) -> Result<()> {
+    let mut doc = Pdf::create(&output_file)?;
 
-use span::Span;
-use sectioner::Sectioner;
-use pages::Pages;
-
-#[derive(Debug, StructOpt)]
-struct Cli {
-    /// The markdown file to read and render
-    markdown_file: String,
-    /// Where to save the generated `.pdf` to
-    #[structopt(long="out", short="o")]
-    output_file: String,
-
-    #[structopt(flatten)]
-    verbosity: Verbosity,
-}
-
-main!(|args: Cli, log_level: verbosity| {
-    let mut doc = Pdf::create(&args.output_file)?;
-
-    let mut markdown_file = File::open(args.markdown_file)?;
+    let mut markdown_file = File::open(markdown_file)?;
     let mut markdown = String::new();
     markdown_file.read_to_string(&mut markdown)?;
 
@@ -110,5 +97,5 @@ main!(|args: Cli, log_level: verbosity| {
     }
 
     doc.finish()?;
-});
-
+    Ok(())
+}
